Rename '/sections' folder in /Download to '/components' for consistency.
